Create SitePhotoManager.vue
