+++ src/components/Site/SiteForm.vue
@@ -123,67 +123,7 @@
             </v-row>
             <v-row>
               <v-col>
-                <h6 class="text-h6 mb-4 mt-7">Add Photos</h6>
-                <v-card-text
-                  id="drop-area"
-                  @dragover.prevent
-                  @drop="handleDrop"
-                  class="drop-area text-subtitle-2 text-medium-emphasis d-flex mb-6"
-                >
-                  <v-icon class="mr-1">mdi-paperclip</v-icon>
-                  Drag and drop your photos here, or
-                  <span @click="triggerFileInput" class="ml-1 add-link"
-                    >click to upload</span
-                  >
-
-                  <input
-                    type="file"
-                    ref="fileInput"
-                    id="fileInput"
-                    multiple
-                    @change="
-                      previewPhotos(($event.target as HTMLInputElement).files)
-                    "
-                    accept="image/jpeg, image/png"
-                    style="display: none"
-                  />
-                </v-card-text>
-
-                <div class="photo-container">
-                  <div
-                    v-if="thingId && photos[thingId]"
-                    v-for="photo in photos[thingId]"
-                    :key="photo.id"
-                    class="photo-wrapper"
-                  >
-                    <img
-                      v-if="!photosToDelete.includes(photo.id)"
-                      :src="photo.link"
-                      class="photo"
-                    />
-                    <v-icon
-                      v-if="!photosToDelete.includes(photo.id)"
-                      color="red-darken-1"
-                      class="delete-icon"
-                      @click="removeExistingPhoto(photo.id)"
-                      >mdi-close-circle</v-icon
-                    >
-                  </div>
-
-                  <div
-                    v-for="(photo, index) in previewedPhotos"
-                    :key="index"
-                    class="photo-wrapper"
-                  >
-                    <img :src="photo" class="photo" />
-                    <v-icon
-                      color="red-darken-1"
-                      class="delete-icon"
-                      @click="removePhoto(index)"
-                      >mdi-close-circle</v-icon
-                    >
-                  </div>
-                </div>
+                <SitePhotoManager :thing-id="thingId" />
               </v-col>
             </v-row>
           </v-col>
@@ -202,29 +142,23 @@
 import { onMounted, reactive, ref, watch } from 'vue'
 import GoogleMap from '../GoogleMap.vue'
 import { useThingStore } from '@/store/thing'
-import { usePhotosStore } from '@/store/photos'
 import { Thing } from '@/types'
 import { siteTypes } from '@/vocabularies'
 import { VForm } from 'vuetify/components'
 import { rules } from '@/utils/rules'
-import Notification from '@/utils/notifications'
 import { storeToRefs } from 'pinia'
 import { api } from '@/services/api'
+import SitePhotoManager from '@/components/Site/SitePhotoManager.vue'
+import { usePhotosStore } from '@/store/photos'
 
 const { thing: storedThing } = storeToRefs(useThingStore())
-const { updatePhotos, fetchPhotos } = usePhotosStore()
-const { photos } = storeToRefs(usePhotosStore())
+const { updatePhotos } = usePhotosStore()
 
 const props = defineProps({ thingId: String })
 const emit = defineEmits(['close'])
 let loaded = ref(false)
 const valid = ref(false)
 const myForm = ref<VForm>()
-
-const newPhotos = ref<File[]>([])
-const photosToDelete = ref<string[]>([])
-const previewedPhotos = ref<string[]>([])
-const fileInput = ref<HTMLInputElement | null>(null)
 
 const mapOptions = ref({
   center: { lat: 39, lng: -100 },
@@ -244,52 +178,6 @@ watch(
     }
   }
 )
-
-function handleDrop(e: DragEvent) {
-  e.preventDefault()
-  let files = e.dataTransfer?.files
-  if (files) {
-    let filteredFiles = Array.from(files).filter(
-      (file) => file.type === 'image/jpeg' || file.type === 'image/png'
-    )
-    if (filteredFiles.length > 0) {
-      previewPhotos(filteredFiles)
-    } else {
-      Notification.toast({
-        message: 'only JPEG and PNG images are allowed',
-        type: 'error',
-      })
-    }
-  }
-}
-
-function previewPhotos(files: File[] | FileList | null) {
-  if (files) {
-    Array.from(files).forEach((photo) => {
-      let reader = new FileReader()
-      reader.onload = (e) => {
-        if ((e.target as FileReader).result) {
-          previewedPhotos.value.push((e.target as FileReader).result as string)
-          newPhotos.value.push(photo)
-        }
-      }
-      reader.readAsDataURL(photo)
-    })
-  }
-}
-
-function triggerFileInput() {
-  if (fileInput.value) fileInput.value.click()
-}
-
-function removePhoto(index: number) {
-  previewedPhotos.value.splice(index, 1)
-  newPhotos.value.splice(index, 1)
-}
-
-function removeExistingPhoto(photoId: string) {
-  photosToDelete.value.push(photoId)
-}
 
 async function populateThing() {
   Object.assign(thing, storedThing.value)
@@ -316,15 +204,14 @@ async function uploadThing() {
     if (props.thingId) {
       try {
         storedThing.value = await api.updateThing(thing)
+        await updatePhotos(storedThing.value!.id)
       } catch (error) {
         console.error('Error updating thing', error)
       }
-      await updatePhotos(props.thingId, newPhotos.value, photosToDelete.value)
     } else {
       try {
         storedThing.value = await api.createThing(thing)
-        if (newPhotos.value)
-          await updatePhotos(storedThing.value!.id, newPhotos.value, [])
+        await updatePhotos(storedThing.value!.id)
       } catch (error) {
         console.error('Error creating thing', error)
       }
@@ -343,47 +230,9 @@ function onMapLocationClicked(locationData: Thing) {
 onMounted(async () => {
   if (props.thingId) {
     await populateThing()
-    await fetchPhotos(props.thingId)
     includeDataDisclaimer.value = !!thing.dataDisclaimer
   } else {
     loaded.value = true
   }
 })
 </script>
-
-<style scoped lang="scss">
-.drop-area {
-  border: 2px dashed #ccc;
-}
-
-.add-link {
-  color: blue;
-  text-decoration: underline;
-  cursor: pointer;
-}
-
-.photo-container {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-}
-
-.photo-wrapper {
-  position: relative;
-  margin-right: 20px;
-  width: 6rem;
-  margin-bottom: 20px;
-}
-
-.photo {
-  width: 100px;
-  height: 100px;
-  object-fit: cover;
-}
-
-.delete-icon {
-  position: absolute;
-  top: -20px;
-  right: -20px;
-}
-</style>
