+++ src/store/photos.ts
@@ -4,47 +4,47 @@ import { Photo } from '@/types'
 import { api } from '@/services/api'
 
 export const usePhotosStore = defineStore('photos', () => {
-  const photos = ref<Record<string, Array<Photo>>>({})
+  const photos = ref<Photo[]>([])
+  const newPhotos = ref<File[]>([])
+  const photosToDelete = ref<string[]>([])
   const loading = ref(false)
 
-  const fetchPhotos = async (thingId: string) => {
-    try {
-      photos.value[thingId] = await api.fetchSitePhotos(thingId)
-    } catch (error) {
-      console.error('Error fetching photos from DB', error)
+  const uploadNewPhotos = async (thingId: string) => {
+    if (newPhotos.value.length > 0) {
+      const data = new FormData()
+      newPhotos.value.forEach((photo) => data.append('files', photo))
+      photos.value = await api.uploadSitePhotos(thingId, data)
     }
   }
 
-  const updatePhotos = async (
-    thingId: string,
-    newPhotos: File[],
-    photosToDelete: string[]
-  ) => {
+  const deleteSelectedPhotos = async (thingId: string) => {
+    for (const photoId of photosToDelete.value) {
+      await api.deleteSitePhoto(thingId, photoId)
+    }
+    photos.value = photos.value.filter(
+      (p) => !photosToDelete.value.includes(p.id)
+    )
+  }
+
+  const updatePhotos = async (thingId: string) => {
     try {
       loading.value = true
-      if (newPhotos.length === 0 && photosToDelete.length === 0) return
-      if (newPhotos.length > 0) {
-        const data = new FormData()
-        newPhotos.forEach((photo) => data.append('files', photo))
-        photos.value[thingId] = await api.uploadSitePhotos(thingId, data)
-      }
-      for (const photoId of photosToDelete) {
-        await api.deleteSitePhoto(thingId, photoId)
-      }
-      photos.value[thingId] = photos.value[thingId].filter(
-        (photo) => !photosToDelete.includes(photo.id)
-      )
+      await uploadNewPhotos(thingId)
+      await deleteSelectedPhotos(thingId)
     } catch (error) {
       console.error('Error updating photos', error)
     } finally {
       loading.value = false
+      newPhotos.value = []
+      photosToDelete.value = []
     }
   }
 
   return {
     photos,
+    newPhotos,
+    photosToDelete,
     loading,
-    fetchPhotos,
     updatePhotos,
   }
 })
